\name{logConCIfunctions}
\alias{logConCIfunctions}
\alias{ftilde}
\alias{c2hat}
\alias{rLCD}
\title{Functions that are used by logConCI}
\description{Functions that are used by \code{\link{logConCI}} and are not intended to be called by the user.}
\references{
Azadbakhsh, M., Jankowski, H. and Gao, X. (2014).
Computing confidence intervals for log-concave densities. 
\emph{Comput. Statist. Data Anal.}, to appear.

Baladbaoui, F., Rufibach, K. and Wellner, J. (2009).
Limit distribution theory for maximum likelihood estimation of a log-concave density.
\emph{Ann. Statist.}, \bold{37(3)}, 1299--1331.

Tarn Duong (2012). ks: Kernel smoothing. 
R package version 1.8.10. \url{http://CRAN.R-project.org/package=ks}
}
\author{
Mahdis Azadbakhsh 

Hanna Jankowski, \email{hkj@mathstat.yorku.ca}, \cr \url{http://www.math.yorku.ca/~hkj}
}
\keyword{htest}
\keyword{nonparametric}

