\name{reliability}
\alias{reliability}
\docType{data}
\title{Reliability dataset used to illustrate log-concave density estimation}
\description{Dataset that contains the data analyzed in Duembgen and Rufibach (2009, Figure 2).}
\usage{data(reliability)}
\format{A vector containing the 786 observations analyzed in Duembgen and Rufibach (2009, Figure 2).}
\source{The data was taken from Duembgen and Rufibach (2009).}
\references{
Duembgen, L. and Rufibach, K. (2009)
Maximum likelihood estimation of a log--concave density and its distribution function: basic properties and uniform consistency. 
\emph{Bernoulli}, \bold{15(1)}, 40--68.  
}
\seealso{See the example in \code{\link{logConDens}} for the analysis of this data.}
\keyword{datasets}
